+++ src/dev/flang/ast/FeaturesAndOuter.java
@@ -90,13 +90,18 @@ public class FeaturesAndOuter extends ANY
    */
   AbstractFeature filter(SourcePosition pos, FeatureName name, java.util.function.Predicate<AbstractFeature> isCandidate)
   {
+    AbstractFeature foundChoice = null;
     var match = false;
     var found = new List<AbstractFeature>();
     for (var f : features.entrySet())
       {
         var ff = f.getValue();
         var fn = f.getKey();
-        if (fn.equalsExceptId(name))  /* an exact match, so use it: */
+        if (ff.isChoice() && !ff.isBaseChoice() /* suppress call to choice type (e.g. bool : choice TRUE FALSE), expect for (inheritance) calls to 'choice'*/)
+          {
+            foundChoice = ff;
+          }
+        else if (fn.equalsExceptId(name))  /* an exact match, so use it: */
           {
             if (CHECKS) check
               (Errors.count() > 0 || !match || fn.argCount() == 0);
@@ -114,7 +119,7 @@ public class FeaturesAndOuter extends ANY
       }
     return switch (found.size())
       {
-      case 0 -> null;
+      case 0 -> foundChoice; // in case we found a matching choice, return it to create a proper error ('Cannot call choice feature').
       case 1 -> found.get(0);
       default ->
       {