+++ src/dev/flang/parser/FList.java
@@ -63,7 +63,7 @@ public class FList extends ANY implements Stmnt
   /**
    * The features in this list
    */
-  final List<Feature> _list = new List<Feature>();
+  final List<Feature> _list;
 
 
   /*--------------------------  constructors  ---------------------------*/
@@ -71,118 +71,14 @@ public class FList extends ANY implements Stmnt
 
   /**
    * Constructor
-   *
-   * @param pos the sourcecode position, used for error messages.
-   *
-   * @param v
-   *
-   * @param m
-   *
-   * @param r
-   *
-   * @param qnames
-   *
-   * @param g
-   *
-   * @param a
-   *
-   * @param i
-   *
-   * @param c
-   *
-   * @param p
    */
-  public FList(SourcePosition pos,
-               Visi v,
-               int m,
-               ReturnType r,
-               List<List<String>> qnames,
-               List<Feature> a,
-               List<AbstractCall> i,
-               Contract c,
-               Impl p) {
-    for (List<String> n : qnames)
-      {
-        if (p.kind_ == Impl.Kind.Of)
-          {
-            var ng = new List<AbstractType>();
-            addFeaturesFromBlock(p._code, ng, p);
-            if (i.isEmpty())
-              {
-                AstErrors.featureOfMustInherit(pos, p.pos);
-              }
-            else
-              {
-                var ic = i.getLast();
-                if (!ic.generics().isEmpty())
-                  {
-                    ic.generics().addAll(ng);
-                  }
-                else
-                  {
-                    ((Call)ic).generics = ng;
-                  }
-              }
-            p = new Impl(p.pos, new Block(p.pos, new List<>()), Impl.Kind.Routine);
-          }
-        _list.add(new Feature(pos, v,m,r,n,a,i,c,p));
-      }
+  public FList(List<Feature> l)
+  {
+    _list = l;
   }
 
 
   /*-----------------------------  methods  -----------------------------*/
-
-
-  /**
-   * For a feature declaration of the form
-   *
-   *   xyz : choice of a, b, c.
-   *
-   * add features a, b, c to _list and they types to g.
-   *
-   * @param s the statements containing the feature declarations to be added, in
-   * this case "a, b, c."
-   *
-   * @param g the list of type to be callected, will be added as generic
-   * arguments to 'choice' in this example
-   *
-   * @param p Impl that contains the position of 'of' for error messages.
-   */
-  private void addFeaturesFromBlock(Stmnt s, List<AbstractType> g, Impl p)
-  {
-    if (s instanceof Block b)
-      {
-        b.statements_.forEach(x -> addFeaturesFromBlock(x, g, p));
-      }
-    else if (s instanceof Feature f)
-      {
-        boolean ok = true;
-        if (f._qname.size() > 1)
-          {
-            AstErrors.featureOfMustContainOnlyUnqualifiedNames(f, p.pos);
-            ok = false;
-          }
-        if (!f.generics().list.isEmpty())
-          {
-            AstErrors.featureOfMustNotHaveFormalGenerics(f, p.pos);
-            ok = false;
-          }
-        if (!f.isConstructor())
-          {
-            AstErrors.featureOfMustContainOnlyConstructors(f, p.pos);
-            ok = false;
-          }
-        if (ok)
-          {
-            _list.add(f);
-            g.add(new Type(f.pos(), f.featureName().baseName(), new List<>(), null));
-          }
-      }
-    else
-      {
-        AstErrors.featureOfMustContainOnlyDeclarations(s, p.pos);
-      }
-  }
 
 
   /**