+++ release_notes.md
@@ -1,3 +1,59 @@
+## 2022-05-31: V0.074
+
+- Fuzion language
+
+  - Overloading with type parameters now works as follows: A feature f with
+    m type parameters and n value parameters can be called with m+n actual
+    arguments (m types and n values) or, if the types can be inferred from the
+    value parameters, with just n actual value arguments.
+
+    If an overloaded f with n formal arguments (type or value) exists, a call
+    with n actual arguments with always call this f with n formal arguments.
+
+- fz command
+
+  - bug fixes for #113, #297, #300
+
+- base lib
+
+  - some changes towards using type parameter arguments instead of generics
+    using '<' / '>'
+
+## 2022-05-18: V0.073
+
+- Fuzion language
+
+  - add support for calling type parameters, i.e., in a feature `f(X type)`, it
+    is now possible to call `X` as in `say X`.  The result is an instance of
+    feature `Type`.
+
+  - add support to declare inner feature in type feature, e.g.
+    ```
+      i32.type.blabla is say "this is blabla in i32.type"
+    ```
+  - type features now inherit from the type features of corresponding to the
+    plain features their corresponding plain feature inherits from.
+
+- stdlib
+
+  - added `fract`, `floor`, `ceil` to `float`.
+
+  - fixed `stream.asList` to work if called repeatedly.
+
+  - fixed `print` and `println` to stdout/stderr to print the whole string
+    atomically at once, i.e., no longer split it up into single characters.
+
+  - added `io.stdin`.
+
+  - added `Functions.compose`.
+
+  - added effect `mutate`. A mutable field `f` with initial value `v` can now be
+    created using `f := mut v`
+
+- fum/fuir:
+
+  - add support for type parameters in .fum file
+
 ## 2022-04-23: V0.072
 
 - Front End